Don't attach laeken and data.table in examples
Skip calls to `base::library` in example codes. Also, do not create deep
copies when calling `recalib`. This is possible because the input dataset
is already copied when it is sent to `simPop::ipu2`.

However, it is important not to use `attr<-` or `names<-` in that case. For that
reason, some additions from 53016eb and 5c3a779 have been revisited