Add aggregateReplicates() and corresponding estimators
Add new functions which applies a given estimator function to all
bootstrap weights and returns a matrix of the form

Group |  w  | w1  | w2  | w3  | w4  | w5
------|-----|-----|-----|-----|-----|-----
G1    | 2.1 | 2.2 | 2.1 | 2.3 | 2.7 | 2.9
G2    | 1.6 | 1.5 | 1.4 | 1.6 | 1.7 | 2.0

Where group is a grouping variable like the group variable in
`calc.stError` and the columns w, w1, ..., w5 denote are the
estimator and the bootstrap replicates belonging to weights
of the same name.

The idea here is to apply the actual estimate of dispersion
(currently standard deviation) to the replicates in a
subsequent call.